pyproject.toml
[project]
name = "snapmark2"
version = "0.1.0"
description = "Desktop screenshot tool with OCR and Markdown generation"
authors = [{name = "User", email = "user@example.com"}]
requires-python = ">=3.8.1"
dependencies = [
    "PyQt5>=5.15.9",
    "mss>=9.0.0",
    "pytesseract>=0.3.10",
    "Pillow>=10.0.0",
    "pynput>=1.7.6",
    "openai>=1.0.0",
    "schedule>=1.2.0",
    "markdown>=3.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
snapmark = "snapmark.main:main"
snapmark-bg = "snapmark.background:run_background_service"

[tool.hatch.build.targets.wheel]
packages = ["snapmark"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]